// src/app/services/catalogue2/catalogue-response.dto.ts
export class CatalogueResponseDto {
  totalPages: number = 0;
  nbJeux: number = 0;
  items: CatalogueItemResponseDto[] = [];
}
export class CatalogueItemResponseDto {

  id: number = 0;

  nom: string = '';

  description: string = '';

  editeur: string = '';

  prix : number = 0;

  prenomVendeur: string = '';

  nomVendeur: string = '';

  image: string = ''; // Add the image attribute

  etat: string = ''; // Add the statut attribute

  // Add other properties as needed
}